feat: add mevblocker product page
