+++ apps/cow-fi/pages/cow-protocol.tsx
@@ -74,7 +74,7 @@ const FAQ_DATA = [
   },
 ]
 
-interface HomeLandingProps {
+interface PageProps {
   siteConfigData: typeof CONFIG
   categories: {
     name: string
@@ -99,7 +99,7 @@ const Wrapper = styled.div`
   gap: 24px;
 `
 
-export default function HomeLanding({ siteConfigData }: HomeLandingProps) {
+export default function Page({ siteConfigData }: PageProps) {
   return (
     <LayoutV2 bgColor={Color.neutral90}>
       <Head>
@@ -658,7 +658,7 @@ export default function HomeLanding({ siteConfigData }: HomeLandingProps) {
           </ContainerCardSection>
         </ContainerCard>
 
-        <ContainerCard bgColor={'transparent'} touchFooter>
+        <ContainerCard bgColor={'transparent'} color={Color.neutral10} touchFooter>
           <ContainerCardSection padding={'0 0 100px'}>
             <SectionTitleWrapper>
               <SectionTitleIcon>
@@ -675,7 +675,7 @@ export default function HomeLanding({ siteConfigData }: HomeLandingProps) {
   )
 }
 
-export const getStaticProps: GetStaticProps<HomeLandingProps> = async () => {
+export const getStaticProps: GetStaticProps<PageProps> = async () => {
   const siteConfigData = CONFIG
   const categoriesResponse = await getCategories()
   const articlesResponse = await getArticles()
