+++ apps/cow-fi/pages/mev-blocker.tsx
@@ -77,7 +77,7 @@ const FAQ_DATA = [
   },
 ]
 
-interface HomeLandingProps {
+interface PageProps {
   siteConfigData: typeof CONFIG
   categories: {
     name: string
@@ -102,7 +102,7 @@ const Wrapper = styled.div`
   gap: 24px;
 `
 
-export default function HomeLanding({ siteConfigData }: HomeLandingProps) {
+export default function Page({ siteConfigData }: PageProps) {
   return (
     <LayoutV2 bgColor={'#3FC4FF'}>
       <Head>
@@ -440,7 +440,7 @@ export default function HomeLanding({ siteConfigData }: HomeLandingProps) {
   )
 }
 
-export const getStaticProps: GetStaticProps<HomeLandingProps> = async () => {
+export const getStaticProps: GetStaticProps<PageProps> = async () => {
   const siteConfigData = CONFIG
   const categoriesResponse = await getCategories()
   const articlesResponse = await getArticles()
